war-premium/requirements.txt
pandas==2.2.0
numpy==1.26.4
scipy==1.12.0
statsmodels==0.14.1
arch==6.3.0
pandas_market_calendars==4.4.1
matplotlib==3.8.2
seaborn==0.13.2
requests
python-dotenv
pyarrow
pytest
yfinance
